refactor: 잔액 필드 타입 변경

+++ back/src/main/java/ssafy/hico/domain/point/dto/ChildApplyTranRequest.java
@@ -2,9 +2,11 @@ package ssafy.hico.domain.point.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ChildApplyTranRequest {
-    private int balance;
-    private double frBalance;
+    private BigDecimal balance;
+    private BigDecimal frBalance;
     private Long countryId;
 }
