+++ back/src/main/java/ssafy/hico/domain/point/repository/FrPointRepository.java
@@ -7,10 +7,12 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import ssafy.hico.domain.point.entity.FrPoint;
 
+import java.math.BigDecimal;
+
 @Repository
 public interface FrPointRepository extends JpaRepository<FrPoint, Long> {
 
     @Modifying
     @Query("UPDATE FrPoint SET balance = :balance WHERE frPointId = :frPointId")
-    void updatePoint(@Param("frPointId") long frPointId, @Param("balance") double balance);
+    void updatePoint(@Param("frPointId") long frPointId, @Param("balance") BigDecimal balance);
 }
