+++ back/src/main/java/ssafy/hico/domain/transaction/dto/response/ChildForeignTransactionResponse.java
@@ -3,16 +3,17 @@ package ssafy.hico.domain.transaction.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 @AllArgsConstructor
 public class ChildForeignTransactionResponse {
     private Long frTranId;
-    private int balance;
+    private BigDecimal balance;
 
     private Long countryId;
-    private double frBalance;
+    private BigDecimal frBalance;
     private String code;
 
     private LocalDateTime createTime;
